% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indodax.R
\name{Indodax}
\alias{Indodax}
\alias{indodax}
\title{API for Trading at Indodax}
\usage{
indodax$new()
indodax$new(url, key, secret)
}
\description{
Connect to Indodax API endpoints
}
\details{
Help you connect to Indodax API endpoints and execute your trading function.

Note: this is not a robot trading. This only helps you connect and execute your trading command.
}
\seealso{
\link{httr} for API management in R
}
\author{
Suberlin Sinaga (2021)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{indodax$new()}}
\item \href{#method-print}{\code{indodax$print()}}
\item \href{#method-posting}{\code{indodax$posting()}}
\item \href{#method-assetInfo}{\code{indodax$assetInfo()}}
\item \href{#method-transHistory}{\code{indodax$transHistory()}}
\item \href{#method-withdrawHistory}{\code{indodax$withdrawHistory()}}
\item \href{#method-depositHistory}{\code{indodax$depositHistory()}}
\item \href{#method-tradeHistory}{\code{indodax$tradeHistory()}}
\item \href{#method-buyHistory}{\code{indodax$buyHistory()}}
\item \href{#method-sellHistory}{\code{indodax$sellHistory()}}
\item \href{#method-trade}{\code{indodax$trade()}}
\item \href{#method-sell}{\code{indodax$sell()}}
\item \href{#method-buy}{\code{indodax$buy()}}
\item \href{#method-openOrders}{\code{indodax$openOrders()}}
\item \href{#method-orderHistory}{\code{indodax$orderHistory()}}
\item \href{#method-getOrder}{\code{indodax$getOrder()}}
\item \href{#method-cancelOrder}{\code{indodax$cancelOrder()}}
\item \href{#method-clone}{\code{indodax$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializing new object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$new(url = NULL, secret = NULL, key = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{indodax api endpoint}

\item{\code{secret}}{indodax secret API}

\item{\code{key}}{indodax key API}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
To initialize this object, you need to call `new()` method
that will span this initialize function. All the arguments
are optional.

When not supplied, you need to store them in you .Renviron as:
      api_key_indodax for key
      secret_key_indodax for secret
      url_api_indodax for url
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Representation to print the object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-posting"></a>}}
\if{latex}{\out{\hypertarget{method-posting}{}}}
\subsection{Method \code{posting()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$posting(payload, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{payload}}{The post data that will be sent}

\item{\code{params}}{The post data in url-encoded form}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Posting the data to the endpoint to get a response
}

\subsection{Returns}{
Fetched object of the Indodax API
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assetInfo"></a>}}
\if{latex}{\out{\hypertarget{method-assetInfo}{}}}
\subsection{Method \code{assetInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$assetInfo(timestamp = 1578304294000, recvWindow = 1578303937000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timestamp}}{the millisecond timestamp of when the request was created and sent}

\item{\code{recvWindow}}{the number of millisecond after timestamp where your request is valid}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Gives us information about our current assets. It will give us the information about the current balance and current held balance
}

\subsection{Returns}{
dataframe of balance and held balance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transHistory"></a>}}
\if{latex}{\out{\hypertarget{method-transHistory}{}}}
\subsection{Method \code{transHistory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$transHistory(
  timestamp = 1578304294000,
  recvWindow = 1578303937000,
  since = "2021-03-28 00:00:00"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timestamp}}{the millisecond timestamp of when the request was created and sent}

\item{\code{recvWindow}}{the number of millisecond after timestamp where your request is valid}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Fetch transaction history from Indodax server
}

\subsection{Returns}{
list of transaction history
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdrawHistory"></a>}}
\if{latex}{\out{\hypertarget{method-withdrawHistory}{}}}
\subsection{Method \code{withdrawHistory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$withdrawHistory()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Gives us the withdrawal history
}

\subsection{Returns}{
Data frame of withdrawal history
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-depositHistory"></a>}}
\if{latex}{\out{\hypertarget{method-depositHistory}{}}}
\subsection{Method \code{depositHistory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$depositHistory()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Gives us the deposit history
}

\subsection{Returns}{
Data frame of deposit history
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tradeHistory"></a>}}
\if{latex}{\out{\hypertarget{method-tradeHistory}{}}}
\subsection{Method \code{tradeHistory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$tradeHistory(
  pairs = "eth_idr",
  count = 1000,
  from_id = 0,
  end_id = 0,
  order = "desc",
  since = "2021-03-28 00:00:00",
  timestamp = 1578304294000,
  recvWindow = 1578303937000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pairs}}{define specific pairs we want to watch. It supports the vector of pairs}

\item{\code{count}}{number of trading history of each interested coin}

\item{\code{from_id}}{trading id we want to watch from}

\item{\code{end_id}}{the last trading id we want to watch}

\item{\code{order}}{the order of the record}

\item{\code{since}}{specifies the interested time frame of the history}

\item{\code{timestamp}}{the millisecond timestamp of when the request was created and sent}

\item{\code{recvWindow}}{the number of millisecond after timestamp where your request is valid}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Gives us the trading history of corresponding coin. We can grab as many as coins we are interested in by passing them as vector.
}

\subsection{Returns}{
Raw content of the fetched API object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buyHistory"></a>}}
\if{latex}{\out{\hypertarget{method-buyHistory}{}}}
\subsection{Method \code{buyHistory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$buyHistory(
  pairs = "eth_idr",
  count = 1000,
  from_id = 0,
  end_id = 0,
  order = "desc",
  since = "2021-03-28 00:00:00",
  timestamp = 1578304294000,
  recvWindow = 1578303937000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pairs}}{define specific pairs we want to watch. It supports the vector of pairs}

\item{\code{count}}{number of trading history of each interested coin}

\item{\code{from_id}}{trading id we want to watch from}

\item{\code{end_id}}{the last trading id we want to watch}

\item{\code{order}}{the order of the record}

\item{\code{since}}{specifies the interested time frame of the history}

\item{\code{timestamp}}{the millisecond timestamp of when the request was created and sent}

\item{\code{recvWindow}}{the number of millisecond after timestamp where your request is valid}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Gives us the buy history of corresponding coin. We can grab as many as coins we are interested in by passing them as vector.
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sellHistory"></a>}}
\if{latex}{\out{\hypertarget{method-sellHistory}{}}}
\subsection{Method \code{sellHistory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$sellHistory(
  pairs = "eth_idr",
  count = 1000,
  from_id = 0,
  end_id = 0,
  order = "desc",
  since = "2021-03-28 00:00:00",
  timestamp = 1578304294000,
  recvWindow = 1578303937000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pairs}}{define specific pairs we want to watch. It supports the vector of pairs}

\item{\code{count}}{number of trading history of each interested coin}

\item{\code{from_id}}{trading id we want to watch from}

\item{\code{end_id}}{the last trading id we want to watch}

\item{\code{order}}{the order of the record}

\item{\code{since}}{specifies the interested time frame of the history}

\item{\code{timestamp}}{the millisecond timestamp of when the request was created and sent}

\item{\code{recvWindow}}{the number of millisecond after timestamp where your request is valid}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Gives us the buy history of corresponding coin. We can grab as many as coins we are interested in by passing them as vector.
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trade"></a>}}
\if{latex}{\out{\hypertarget{method-trade}{}}}
\subsection{Method \code{trade()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$trade(
  pair,
  type,
  price,
  timestamp = 1578304294000,
  recvWindow = 1578303937000,
  idr = NULL,
  val = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{define specific pairs we want to watch. It supports the vector of pairs}

\item{\code{type}}{type of command to execute (sell or buy)}

\item{\code{price}}{at price should the command will be executed}

\item{\code{timestamp}}{the millisecond timestamp of when the request was created and sent}

\item{\code{recvWindow}}{the number of millisecond after timestamp where your request is valid}

\item{\code{idr}}{amount of rupiah(s) that we want to spend to buy the coin(s)}

\item{\code{val}}{amount of coin(s) that we want to sell}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Function to execute buy and sell. Do not directly execute the function. Use `sell()` or `buy()`.
}

\subsection{Returns}{
raw data from server
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sell"></a>}}
\if{latex}{\out{\hypertarget{method-sell}{}}}
\subsection{Method \code{sell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$sell(pair, val, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{define specific pairs we want to watch. It supports the vector of pairs}

\item{\code{val}}{amount of coin(s) that we want to sell}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Execute sell command.
}

\subsection{Returns}{
data_frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buy"></a>}}
\if{latex}{\out{\hypertarget{method-buy}{}}}
\subsection{Method \code{buy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$buy(pair, idr, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{define specific pairs we want to watch. It supports the vector of pairs}

\item{\code{idr}}{amount of rupiah(s) that we want to spend to buy the coin(s)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Execute buy command.
}

\subsection{Returns}{
data_frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openOrders"></a>}}
\if{latex}{\out{\hypertarget{method-openOrders}{}}}
\subsection{Method \code{openOrders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$openOrders(
  pair = NULL,
  timestamp = 1578304294000,
  recvWindow = 1578303937000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{define specific pairs we want to watch. It supports the vector of pairs}

\item{\code{timestamp}}{the millisecond timestamp of when the request was created and sent}

\item{\code{recvWindow}}{the number of millisecond after timestamp where your request is valid}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Execute command to open order position
}

\subsection{Returns}{
raw content of the function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-orderHistory"></a>}}
\if{latex}{\out{\hypertarget{method-orderHistory}{}}}
\subsection{Method \code{orderHistory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$orderHistory(
  pairs,
  count = 1000,
  from = 0,
  timestamp = 1578304294000,
  recvWindow = 1578303937000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pairs}}{define specific pairs we want to watch. It supports the vector of pairs}

\item{\code{count}}{number of trading history of each interested coin}

\item{\code{from}}{order id we want to watch from}

\item{\code{timestamp}}{the millisecond timestamp of when the request was created and sent}

\item{\code{recvWindow}}{the number of millisecond after timestamp where your request is valid}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Get the details order history opened
}

\subsection{Returns}{
raw content of the function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOrder"></a>}}
\if{latex}{\out{\hypertarget{method-getOrder}{}}}
\subsection{Method \code{getOrder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$getOrder(
  pair,
  order_id,
  timestamp = 1578304294000,
  recvWindow = 1578303937000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{define specific pair we want to watch. It does not support the vector of pair.}

\item{\code{order_id}}{order_id that we want to check the detail of}

\item{\code{timestamp}}{the millisecond timestamp of when the request was created and sent}

\item{\code{recvWindow}}{the number of millisecond after timestamp where your request is valid}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Get the details of a specific order
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cancelOrder"></a>}}
\if{latex}{\out{\hypertarget{method-cancelOrder}{}}}
\subsection{Method \code{cancelOrder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$cancelOrder(
  pair,
  order_id,
  type,
  timestamp = 1578304294000,
  recvWindow = 1578303937000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{define specific pair we want to watch. It does not support the vector of pair.}

\item{\code{order_id}}{order_id that we want to check the detail of}

\item{\code{type}}{order type that we want to cancel (buy or sell)}

\item{\code{timestamp}}{the millisecond timestamp of when the request was created and sent}

\item{\code{recvWindow}}{the number of millisecond after timestamp where your request is valid}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Cancel specific order id
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{indodax$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
